codebox/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
Django==3.0.3
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-richtextfield==1.5.0
idna==2.10
Pillow==7.1.2
pytz==2020.5
requests==2.23.0
sqlparse==0.4.1
urllib3==1.25.11